backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
motor>=3.4.0
beanie>=1.26.0
python-dotenv>=1.0.0
openai>=1.30.0
reportlab>=4.0.0
python-multipart>=0.0.9
pydantic[email]>=2.5.0
httpx>=0.27.0
